requirements.txt
fastapi==0.91.0
uvicorn==0.20.0
requests==2.28.2
cryptography==39.0.1
python-qbittorrent==0.4.3